+++ src/main/java/com/epam/test/automation/java/practice6/SalesPerson.java
@@ -1,10 +1,12 @@
 package com.epam.test.automation.java.practice6;
 
+import java.math.BigDecimal;
+
 public class SalesPerson extends Employee {
 
     private double percent;
 
-    public SalesPerson(String name, double salary, double bonus) {
+    public SalesPerson(String name, BigDecimal salary, double bonus) {
         super(name, salary);
         this.percent = bonus;
     }
